+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   of any python project.
 - Throw a warning if a `Task` class instance received unknown inputs
 
+### Fixed
+
+- Error was logged at INFO level in app logs and ewoks events. Now error is logged at the error level.
+
 ## [5.0.0rc1] - 2026-04-13
 
 ### Added